+++ src/Bozboz/Admin/Controllers/PermissionAdminController.php
@@ -3,7 +3,6 @@
 namespace Bozboz\Admin\Controllers;
 
 use Bozboz\Admin\Permissions\PermissionAdminDecorator;
-use Illuminate\Support\Facades\View;
 
 class PermissionAdminController extends ModelAdminController
 {