+++ src/permissions.php
@@ -7,23 +7,23 @@ $permissions->define([
 	'edit_profile'    => 'Bozboz\Admin\Permissions\UserRule',
 
 	'view_anything'   => 'Bozboz\Permissions\Rules\GlobalRule',
-	'create_anything' => 'Bozboz\Permissions\Rules\GlobalRule',
-	'edit_anything'   => 'Bozboz\Permissions\Rules\Rule',
-	'delete_anything' => 'Bozboz\Permissions\Rules\Rule',
+	'create_anything' => 'Bozboz\Permissions\Rules\ModelRule',
+	'edit_anything'   => 'Bozboz\Permissions\Rules\ModelRule',
+	'delete_anything' => 'Bozboz\Permissions\Rules\ModelRule',
 
 	'view_pages'  => 'Bozboz\Permissions\Rules\GlobalRule',
-	'create_page' => 'Bozboz\Permissions\Rules\GlobalRule',
-	'edit_page'   => 'Bozboz\Permissions\Rules\Rule',
-	'delete_page' => 'Bozboz\Permissions\Rules\Rule',
+	'create_page' => 'Bozboz\Permissions\Rules\ModelRule',
+	'edit_page'   => 'Bozboz\Permissions\Rules\ModelRule',
+	'delete_page' => 'Bozboz\Permissions\Rules\ModelRule',
 
 	'view_users'   => 'Bozboz\Permissions\Rules\GlobalRule',
-	'create_user'  => 'Bozboz\Permissions\Rules\GlobalRule',
-	'edit_user'    => 'Bozboz\Permissions\Rules\Rule',
-	'delete_user'  => 'Bozboz\Permissions\Rules\Rule',
+	'create_user'  => 'Bozboz\Permissions\Rules\ModelRule',
+	'edit_user'    => 'Bozboz\Permissions\Rules\ModelRule',
+	'delete_user'  => 'Bozboz\Permissions\Rules\ModelRule',
 
 	'view_media'   => 'Bozboz\Permissions\Rules\GlobalRule',
-	'create_media' => 'Bozboz\Permissions\Rules\GlobalRule',
-	'edit_media'   => 'Bozboz\Permissions\Rules\Rule',
-	'delete_media' => 'Bozboz\Permissions\Rules\Rule',
+	'create_media' => 'Bozboz\Permissions\Rules\ModelRule',
+	'edit_media'   => 'Bozboz\Permissions\Rules\ModelRule',
+	'delete_media' => 'Bozboz\Permissions\Rules\ModelRule',
 
 ]);