+++ src/views/media/overview.blade.php
@@ -4,7 +4,7 @@
 	<ul class="js-mason secret-list media-view">
 	@foreach ($report->getRows() as $row)
 		<li class="masonry-item">
-			@if ($canEdit($row->getId()))
+			@if ($row->check($canEdit))
 				<a href="{{ URL::action($editAction, array($row->getId())) }}">
 					{{ $row->getColumn('image') }}
 				</a>
@@ -14,7 +14,7 @@
 			<div class="icons">
 				<p>{{ $row->getColumn('caption') }}</p>
 
-				@if ($canDelete($row->getId()))
+				@if ($row->check($canDelete))
 					{{ Form::open(['action' => array($destroyAction, $row->getId()), 'method' => 'DELETE']) }}
 						<button data-warn="true" class="btn btn-danger btn-xs" type="submit">
 							<i class="fa fa-minus-square"></i>