+++ src/views/partials/nav.blade.php
@@ -54,7 +54,7 @@
 				<li class="dropdown">
 					<a href="#" class="dropdown-toggle" data-toggle="dropdown"><i class="fa fa-cog"></i>Logged in as {{ $user->first_name }} <b class="caret"></b></a>
 					<ul class="dropdown-menu">
-						@if (Bozboz\Permissions\RuleStack::with('edit_profile', $user->id)->then('edit_anything')->isAllowed())
+						@if (Bozboz\Permissions\RuleStack::with('edit_profile', $user)->then('edit_anything')->isAllowed())
 							<li><a href="/admin/users/{{ $user->id }}/edit"><i class="fa fa-wrench"></i> Edit Profile</a></li>
 						@endif
 						<li><a href="/"><i class="fa fa-desktop"></i> View Website</a></li>